service unavailable exception added

+++ jimena/jimena-web-layer/src/test/java/net/tecgurus/web/layer/actions/MockSignUpProcessor.java
@@ -5,7 +5,7 @@ import net.tecgurus.business.layer.SignUpProcessor;
 public class MockSignUpProcessor implements SignUpProcessor {
 
 	@Override
-	public int perform(String userName, String userEmail,
+	public long perform(String userName, String userEmail,
 			String userPassword, String userConfirmPass) {
 		if(userName != null){
 			return 1;
